Fix pgsql_fdw to create tuplestore in Begin.
If SAVEPOINT is set between DECLARE and first FETCH, and user rolls back
to it after first FETCH, tuplestore is released undesirably in cleanup
of the subtransaction which started at SAVEPOINT.

We need to create tuplestore in the context of the transaction in which
the query was initiated, so we moved tuplestore_begin_heap from
IterateForeignScan to BeginForeignScan.  This causes little overhead,
but it's limited in the case when the scan is innter scan of a nested
loop which has 0-row scan on outer, so such overhead seems ignorable.

cleanup_connection is also fixed to skip closing connection if the
cleanup is for a sub-transaction.  This is required for queries executed
after SAVEPOINT recovery.