Enable to build in server where oracle client is installed.
Though Makefile sets PG_CPPFLAGS to the path of header files of
Oracle, it assumes that Oracle Server is installed. But, some
users maybe want to build when Oracle Database Client is installed.
In that case, the location of oracle header files is different.
With this fix, PG_CPPFLAGS also check the location where the
environmental variable ORACLE_INCLUDE specifies.

The original patch was written by Sachin Kotwal, and modified by me.

    The original patch was written by Sachin Kotwal, and modified by me.